+++ container-engine-lib/lib/backend_interface/objects/service/service_status.go
@@ -1,0 +1,10 @@
+package service
+
+//go:generate go run github.com/dmarkham/enumer -trimprefix=UserServiceStatus_ -transform=snake-upper -type=UserServiceStatus
+type UserServiceStatus int
+const (
+	UserServiceStatus_Registered UserServiceStatus = iota
+	UserServiceStatus_Running
+	UserServiceStatus_Stopped	// Indicates that the Service has been stopped and can no longer be used
+)
+