+++ container-engine-lib/docs/changelog.md
@@ -4,6 +4,14 @@
 
 ### Fixes
 * Fix an issue with network container IPs not being correctly made available when starting a DockerKurtosisBackend in API container mode
+* Allowed removal of user service registrations independent of the underlying service registration (which is necessary for deregistering services in the API container)
+
+### Breaking Changes
+* `user_service_registration.ServiceID` is now once again `service.ServiceID`
+    * Users should update their code
+* `UserServiceRegistration` objects have been removed
+    * Users should migrate to using the new migration flow
+* 
 
 # 0.21.1
 ### Fixes